fix: remove redundant type casts from InputUtils method calls
InputUtils methods already have proper return type declarations:
- filterInt() -> int
- filterFloat() -> float
- filterDate() -> string

Remove unnecessary casts that were redundant with method signatures:
- (int) InputUtils::filterInt(...) -> InputUtils::filterInt(...)
- (float) InputUtils::filterFloat(...) -> InputUtils::filterFloat(...)
- (string) InputUtils::filterDate(...) -> InputUtils::filterDate(...)

Impact: 12 files, 20 redundant casts removed
All PHP syntax validation passed ✓

+++ src/CSVCreateFile.php
@@ -87,7 +87,7 @@ if ($sSource === 'cart') {
     if (!empty($_POST['Classification'])) {
         $count = 0;
         foreach ($_POST['Classification'] as $Cls) {
-            $Class[$count++] = (int) InputUtils::filterInt($Cls);
+            $Class[$count++] = InputUtils::filterInt($Cls);
         }
         if ($count === 1) {
             if ($Class[0]) {
@@ -105,7 +105,7 @@ if ($sSource === 'cart') {
     if (!empty($_POST['FamilyRole'])) {
         $count = 0;
         foreach ($_POST['FamilyRole'] as $Fmr) {
-            $Class[$count++] = (int) InputUtils::filterInt($Fmr);
+            $Class[$count++] = InputUtils::filterInt($Fmr);
         }
         if ($count === 1) {
             if ($Class[0]) {
@@ -122,14 +122,14 @@ if ($sSource === 'cart') {
 
     if (!empty($_POST['Gender'])) {
         $sWhereExt .= 'AND per_Gender = ? ';
-        $whereParams[] = (int) InputUtils::filterInt($_POST['Gender']);
+        $whereParams[] = InputUtils::filterInt($_POST['Gender']);
         $whereTypes .= 'i';
     }
 
     if (!empty($_POST['GroupID'])) {
         $count = 0;
         foreach ($_POST['GroupID'] as $Grp) {
-            $Class[$count++] = (int) InputUtils::filterInt($Grp);
+            $Class[$count++] = InputUtils::filterInt($Grp);
         }
         if ($count === 1) {
             if ($Class[0]) {
