+++ src/fundraiser/routes/paddle-num.php
@@ -185,7 +185,7 @@ $app->post('/{fundraiserId}/paddle-numbers/editor[/{paddleId}]', function (Reque
 
     $_SESSION['iCurrentFundraiser'] = $fundraiserId;
 
-    $iNum   = (int) InputUtils::filterInt($body['Num'] ?? '0');
+    $iNum   = InputUtils::filterInt($body['Num'] ?? '0');
     $iPerID = (int) InputUtils::legacyFilterInput($body['PerID'] ?? '0');
 
     // On reassignment: delete the old owner's multibuy rows so their statement
